move rendering code to cli/render.py
