+++ harness/determined/cli/render.py
@@ -6,13 +6,12 @@ import os
 import pathlib
 import sys
 from datetime import timezone
-from typing import Any, Dict, Iterable, List, Optional, Sequence, Union
+from typing import Any, Dict, Iterable, List, Optional, Sequence, TextIO, Union
 
 import dateutil.parser
 import tabulate
 import termcolor
 
-from determined import cli
 from determined import util as det_util
 from determined.common import util, yaml
 
@@ -189,6 +188,84 @@ def yes_or_no(prompt: str) -> bool:
         return False
 
 
+KEY = "blue"
+PRIMITIVES = None  # avoid coloring to common fg or bg colors.
+SEPARATORS = "yellow"
+STRING = "green"
+
+
+def render_colorized_json(
+    obj: Any, out: TextIO, indent: str = "  ", sort_keys: bool = False
+) -> None:
+    """
+    Render JSON object to output stream with color.
+    """
+
+    def do_render(obj: Any, depth: int = 0) -> None:
+        if obj is None:
+            out.write(termcolor.colored("null", PRIMITIVES))
+            return
+
+        if isinstance(obj, bool):
+            out.write(termcolor.colored(str(obj).lower(), PRIMITIVES))
+            return
+
+        if isinstance(obj, str):
+            out.write(termcolor.colored(json.dumps(obj), STRING))
+            return
+
+        if isinstance(obj, (int, float)):
+            out.write(termcolor.colored(str(obj), PRIMITIVES))
+            return
+
+        if isinstance(obj, (list, tuple)):
+            if len(obj) == 0:
+                out.write(termcolor.colored("[]", SEPARATORS))
+                return
+
+            out.write(termcolor.colored("[", SEPARATORS))
+            first = True
+            for item in obj:
+                if not first:
+                    out.write(termcolor.colored(",", SEPARATORS))
+                first = False
+                out.write("\n")
+                out.write(indent * (depth + 1))
+                do_render(item, depth + 1)
+            out.write("\n")
+            out.write(indent * depth)
+            out.write(termcolor.colored("]", SEPARATORS))
+            return
+
+        if isinstance(obj, dict):
+            if len(obj) == 0:
+                out.write(termcolor.colored("{}", SEPARATORS))
+                return
+
+            out.write(termcolor.colored("{", SEPARATORS))
+            first = True
+            keys = sorted(obj.keys()) if sort_keys else obj.keys()
+            for key in keys:
+                value = obj[key]
+                if not first:
+                    out.write(termcolor.colored(",", SEPARATORS))
+                first = False
+                out.write("\n")
+                out.write(indent * (depth + 1))
+                out.write(termcolor.colored(json.dumps(key), KEY))
+                out.write(termcolor.colored(": ", SEPARATORS))
+                do_render(value, depth + 1)
+            out.write("\n")
+            out.write(indent * depth)
+            out.write(termcolor.colored("}", SEPARATORS))
+            return
+
+        raise ValueError(f"unsupported type: {type(obj).__name__}")
+
+    do_render(obj, depth=0)
+    out.write("\n")
+
+
 def _coloring_enabled() -> bool:
     return sys.stdout.isatty() and os.environ.get("DET_CLI_COLORIZE", "").lower() in ("1", "true")
 
@@ -201,7 +278,7 @@ def print_json(data: Union[str, Any]) -> None:
         if isinstance(data, str):
             data = json.loads(data)
         if _coloring_enabled():
-            cli.render_json(data, sys.stdout, indent="  ", sort_keys=True)
+            render_colorized_json(data, sys.stdout, indent="  ", sort_keys=True)
         else:
             formatted_json = det_util.json_encode(data, sort_keys=True, indent="  ")
             print(formatted_json)
